Function reqchan::channel

pub fn channel<T>() -> (Requester<T>, Responder<T>)

This function creates a reqchan and returns a tuple containing the two ends of this bidirectional request->response channel.

Example

extern crate reqchan;

#[allow(unused_variables)]
let (requester, responder) = reqchan::channel::<u32>();